// frontend/src/app/models/settings/user-settings.model.ts
export interface UserSettings {
    userId: number;
	profilePrivacyMode: string;
	ethnicityPrivacyMode: string;
	religionPrivacyMode: string;
	politicalPrivacyMode: string;
	sexualOrientationPrivacyMode: string;
	pronounsPrivacyMode: string;
	birthdayPrivacyMode: string;
	cityStatePrivacyMode: string;
	emailPrivacyMode: string;
	phonePrivacyMode: string;
	defaultActivityPrivacyMode: string;
	darkMode: boolean;
	hasPreschoolKids: boolean;
	hasElementaryKids: boolean;
	hasMiddleSchoolKids: boolean;
	hasHighSchoolKids: boolean;
	hasCollegeKids: boolean;
	hasKidsPrivacyMode: string;
	insertDate: Date;
	updateDate: Date;
}